ARROW-13544 [Java]: Remove APIs that have been deprecated for long (Changes to JDBC)
